Rename vote_encoder to point_encoding
Vote encoder is too generic since it could cover high level vote encoding as well
