+++ lib/av_client/encoding/point_encoding.ts
